backend/function/requirements.txt
Pillow==11.1.0
google-cloud-storage==2.19.0
